Fix/navigation

+++ presentation/viewModel/src/main/kotlin/com/berlin/aflami/viewmodel/searchcountry/SearchByCountryEffect.kt
@@ -2,5 +2,5 @@ package com.berlin.aflami.viewmodel.searchcountry
 
 sealed class SearchByCountryEffect {
     object NavigatedBack : SearchByCountryEffect()
-    data class NavigatedToMovieDetailsScreen(val movieId: Int, val mediaType: String): SearchByCountryEffect()
+    data class NavigatedToMovieDetailsScreen(val movieId: Long, val mediaType: String): SearchByCountryEffect()
 }
